Change function to delete users in lectures updated

+++ app/Domains/Support/helpers.php
@@ -48,12 +48,14 @@ if (!function_exists('deleteUsersLecturesByLectureId')) {
     function deleteUsersLecturesByLectureId($id)
     {
         $users_lectures = UsersLecture::where('lecture_id', '=', $id)->get();
-        foreach ($users_lectures as $user_lecture) {
-            $roles = LectureUserRole::where('user_lecture_id', '=', $users_lectures->id)->get();
-            foreach ($roles as $role) {
-                $role->delete();
+        if (count($users_lectures)) {
+            foreach ($users_lectures as $user_lecture) {
+                $roles = LectureUserRole::where('user_lecture_id', '=', $user_lecture->id)->get();
+                foreach ($roles as $role) {
+                    $role->delete();
+                }
+                $user_lecture->delete();
             }
-            $user_lecture->delete();
         }
     }
 }
